feat: getPrizeMoney 메서드의 반환 타입을 long에서 int로 변경

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -45,7 +45,7 @@ public enum LottoPrize {
         return describePrize.formatted(matchCount, "", formattedPrizeMoney);
     }
 
-    public long getPrizeMoney() {
+    public int getPrizeMoney() {
         return prizeMoney;
     }
 }
